Add first pass at filter docs for API #388
This commit adds a first pass at trying to add options for filtering to
the /api/v0/databases/<id>/types/ output.

The idea here is to make it pretty trivial to add the options for new
types to the endpoint. I'm not totally sold on the this being the best
format per-se, but I like how it gets injected automatically for a given
type and depends only on providing a definition. The definition can
change, but the place they go stays constant.