Add svelte-ignore directive
Svelte linter cannot differentiate between static/pure and stateful functions
so it warns when using module context functions in reactive blocks.

The ability to disable this specific warning will be available when svelte is
updated to version 3.39.0 or above.

This commit includes the currently ineffective directive that will take
effect if above mentioned update is performed.