Remove "section" language from display options
Previously, the UI for the "Filter", "Sort", and "Group" display
options was all in one pane, split into sections -- so the
`<section>` tag was appropriate, as well as the word "Section" in each
component name. Now that we're moving each display option into its
own Dropdown, the "Section" language may be confusing.