Initial implementation of frontend filtering by duplicates
The initial implementation of creating a frontend filter to filter by duplicates has been implemented here.
Because of the filtering by duplicates being a special case, the user is not allowed to have any filters past the filter by duplicates filter.
The code has also been updated to follow this behavior. Because this filtering is a special case, it has special handling in meta.ts which
specifically checks for this filtering behavior.