Limit Frontend Handling of get_duplicates op in frontend
The backend currently has a bug where it handles the get_duplicates
operation in a way that is incompatible with the rest of the filters.
This commit seeks to limit the special handling for this bug
to just a single area in the code, so that once the bug is fixed,
it will be easy to undo the special handling.