[#869] added type_options support, some relevant tests
This commit updates the structure of `MASTER_DB_TYPE_MAP_SPEC`, each `VALID` tuple
can now accept an optional dictionary that represents `type_options` to be
provided to the target type.

For instance, a `URI` type "https://centerofci.org" can be cast to a `CHAR` type
with `type_options = {"length": 22}`. If no type options are provided, length
is assumed to be 1. This type cast test can be specified as the`VALID` tuple

`("https://centerofci.org", "https://centerofci.org", {"length": 22})`