Use a fixed date for Date Format option labels
Before:

For example, if someone in Canada is using the app on 2022-05-05, the
"Date Format" options for "US" and "European" would be identical and
the user might not understand which option would give them the format
they want.

After:

2001-12-31 is used as an example date, guaranteeing that the month and
day digits will be immediately distinctive from one another regardless
of today's date.