Improve first time setup blocking logic
Previous blocking logic waited until the container reported itself as
ready, which isn't meaningful in this case and led to race conditions.
These changes make the script wait until pip dependencies are installed
and postgres is ready. My limited testing leads me to believe that
that's enough to know if we can run migrations and the installation
script.