Move logic fixing array_length into ArrayLength
We don't need a DBFunction like Coalesce at the moment; and the
motivation behind introducing that logic right now is that we're
fixing an edge case of ArrayLength, so we'll probably prefer to
do that inside ArrayLength itself.

Also, notice that I broke lines to keep code width not more than
80 characters. We don't have hard rules for this yet, but this is
a good guideline to keep. Also, I personally limit docstring and
comment to a higher width to 100 characters.