Fix TS error in Tree component
Note: this change results in a loss of functionality. I couldn't figure
out what the heck was causing this error. So strange! But I didn't want
to spend too much time on it. And this component is not used anywhere
currently. Even if we do use it somewhere at some point, I'm not sure
that we'll actually need to customize the empty behavior. So I'm kicking
this can down the road.