Add Join button to table page actions pane
Cursor prompts:

The table view has an actions pane, implemented in @ActionsPane.svelte.
It has dropdowns for "Filter", "Sort", and "Group". Add a new dropdown
button, after "Group". The button label should be "Join" (and should be
translated through the i18n system). The icon should be the same link
icon used for a foreign key column. When the dropdown opens, it should
display the text "TODO" (untranslated).

---

Get rid of Join.svelte. Just inline this content for now.

---

Add an entry for the string 'join' in the i18n system. Just add it to
the English language file.